Move renderer monoid to its own module.
